fix: directory cleanup, codebase cleanup
